+++ java/shared/java/org/signal/libsignal/usernames/Username.java
@@ -54,7 +54,7 @@ public final class Username {
 
   public Username(String username) throws BaseUsernameException {
     this.username = Objects.requireNonNull(username, "username");
-    this.hash = hash(username);
+    this.hash = filterExceptions(BaseUsernameException.class, () -> Native.Username_Hash(username));
   }
 
   private Username(String username, byte[] hash) {
@@ -131,28 +131,6 @@ public final class Username {
     final byte[] entropy = Arrays.copyOfRange(bytes, 0, 32);
     final byte[] enctyptedUsername = Arrays.copyOfRange(bytes, 32, bytes.length);
     return new UsernameLink(entropy, enctyptedUsername);
-  }
-
-  @Deprecated
-  public static List<String> generateCandidates(
-      String nickname, int minNicknameLength, int maxNicknameLength) throws BaseUsernameException {
-    Object[] names =
-        filterExceptions(
-            BaseUsernameException.class,
-            () -> Native.Username_CandidatesFrom(nickname, minNicknameLength, maxNicknameLength));
-    return Arrays.asList((String[]) names);
-  }
-
-  @Deprecated
-  public static byte[] hash(String username) throws BaseUsernameException {
-    return filterExceptions(BaseUsernameException.class, () -> Native.Username_Hash(username));
-  }
-
-  @Deprecated
-  public static byte[] generateProof(String username, byte[] randomness)
-      throws BaseUsernameException {
-    return filterExceptions(
-        BaseUsernameException.class, () -> Native.Username_Proof(username, randomness));
   }
 
   public static void verifyProof(byte[] proof, byte[] hash) throws BaseUsernameException {